+++ docs/CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+#### Device API
+
+- Check for firmware updates
+
 ### Fixed
 
 - Get network overview now also works synchroniously